mcp_servers/easyocr-mcp/requirements.txt
# EasyOCR MCP Server Requirements
# Optimized for Low-Resource Usage on Asus x515ae

# Core dependencies
easyocr>=1.7.0
Pillow>=10.0.0
numpy>=1.24.0

# MCP dependencies
mcp>=1.0.0
fastmcp>=0.1.0

# Error handling and logging
colorlog>=6.7.0

# Optional dependencies for enhanced functionality
opencv-python>=4.8.0  # For advanced image processing
scikit-image>=0.21.0  # For additional image processing

# Development dependencies (optional)
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0